Support decimal values and SDK user agents (#4)
## Summary\n\n- preserve fractional price, cost, total-cost, and credit
values with the native decimal-capable type for each language\n- send
{SDK Name}/{SDK Version} as the default User-Agent, with a client
configuration override\n- bump the SDK version to 0.2.0\n\n##
Verification\n\n- formatting and lint targets pass\n- build and package
checks pass\n- request-capture and fractional-value smoke checks pass

+++ src/Adsefid.Sdk/Messenger/Models/SendTemplateMessenger.cs
@@ -55,7 +55,7 @@ public sealed class SendTemplateMessengerResponse
     public DateTimeOffset? ExpiryDate { get; init; }
 
     [JsonPropertyName("cost")]
-    public required long Cost { get; init; }
+    public required decimal Cost { get; init; }
 
     [JsonPropertyName("receptor")]
     public required string Receptor { get; init; }
